+++ Assets/Resources/QuestPrefab/Test_MouseSubjugation.prefab
@@ -46,4 +46,4 @@ MonoBehaviour:
   serializedNorma:
     list:
     - Key: Mouse
-      Value: 5
+      Value: 2
