--- Assets/Scripts/Quest/Quest_Behaviour.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using System;
-
-enum QuestCategory
-{
-    Main,sub
-}
-
-public class Quest_Behaviour:MonoBehaviour
-{
-    QuestCategory questCategory;
-    public string questName { get; private set; }
-    string questTarget;
-    int? deadline;
-    string queriedBy;
-    public Action<GameObject> WhenQuestAccepted { get; private set; }
-    public Action<GameObject> WhenQuestCleared { get; private set; }
-    public Action<GameObject> WhenQuestFailed { get; private set;}
-}
