+++ Assets/Scripts/Quest_General/Quest_Behaviour.cs
@@ -1,0 +1,25 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using System;
+
+enum QuestCategory
+{
+    Main,sub
+}
+
+public class Quest_Behaviour:MonoBehaviour
+{
+    QuestCategory questCategory;
+    public string questName { get; private set; }
+    string questTarget;
+    int? deadline;
+    string queriedBy;
+    public Action WhenQuestAccepted { get; protected set; }
+    public Action WhenQuestCleared { get; protected set; }
+    public Action WhenQuestFailed { get; protected set;}
+
+    virtual protected void Start() {
+        questName = this.gameObject.name;//ゲームオブジェクトの名前と同じでいいならこれ　check
+    }
+}
