+++ Assets/Scripts/Quest_Individual.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0a23911725858564a815d769b9c88a48
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
